+++ RELEASE_NOTES.md
@@ -1,5 +1,6 @@
 # v0.13.4
 xxxx-yy-zz
+* Codegen change to let rustc derive `Default` in most cases
 * API spec update 2021-10-06, 2021-10-21, 2021-11-03, 2021-11-17
     * files::upload_session_finish_batch is deprecated by a _v2 version.
     * app auth added for files::list_folder and files::list_folder_continue